% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coverage.R
\name{check_coverage}
\alias{check_coverage}
\title{Check Coverage}
\usage{
check_coverage(DB, females, males, GT = 40)
}
\arguments{
\item{DB}{RSQLite database generated by \code{make_database}}

\item{females}{Sample IDs of the female to compare to a male. The Sample ID
must match the ones found in the \code{DB}.}

\item{males}{Sample IDs of the males to compare to a female. The Sample ID
must match the ones found in the \code{DB}.}

\item{GT}{percent of genotypes you want to}
}
\value{
dataframe or message
}
\description{
Check the percentage of genotyped loci with coverage. \cr
Returns list of samples under specified percentage.
}
\examples{
\dontrun{
check_coverage(DBs, females, males, 40)
}
}
