% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_rankings.R
\name{get_all_rankings}
\alias{get_all_rankings}
\title{Rank Mates}
\usage{
get_all_rankings(
  DB,
  females,
  males,
  type = "all_alleles",
  bonus = NULL,
  weighted_alleles = NULL
)
}
\arguments{
\item{DB}{RSQLite database generated by \code{make_database}}

\item{females}{Vector of sample IDs of the females to compare to the males. The Sample IDs
must match the ones found in the \code{DB}.}

\item{males}{Vector of sample IDs of the males to compare to the females. The Sample IDs
must match the ones found in the \code{DB}.}

\item{type}{The type of mating that is advantageous.
\enumerate{
\item \code{all_alleles}: the user specifies for each locus what is advantageous.
\item \code{assortive}: Assortive (same alleles) is advantageous for all loci
\item \code{disassortive}: Disassortive (different alleles) is advantageous for all loci
}}

\item{bonus}{Numeric value to increase mating score if the state of alleles
specified in \code{weighted_alleles} are met. If you specify .15 as your bonus,
and 2/3 of your loci match the advantageous, then .10 will be used as the bonus.
The final score will be increased by 10\%. Must specify \code{weighted_alleles}.}

\item{weighted_alleles}{vector listing allele IDs to calculate the bonus. Must
specify \code{bonus}.}
}
\value{
List of dataframes - dataframe for each female ranked by best mating
score.
}
\description{
Wrapper function to score the mating of all females to all males. \cr
This function is a wrapper function for \code{rank_mates}. \cr
It is not recommended to use more than 10x10 samples.
}
\examples{
\dontrun{
all_mates <- get_all_rankings(DB = DBs, female = females,
                              males = males, type = "all_alleles",
                              bonus=NULL, weighted_alleles=NULL)
}
}
