% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_database.R
\name{make_database}
\alias{make_database}
\title{Make Database}
\usage{
make_database(gtseq, metadata, allele_info, verbose = TRUE)
}
\arguments{
\item{gtseq}{Comma separated file that is the output of the GT-seq pipeline.
Sample IDs must match the Sample IDs in the \code{metadata} CSV file. The Allele IDs
from the column names must match the \code{site_id} found in the \code{allele_info} CSV file.}

\item{metadata}{Comma separated file that includes two columns: \code{Sample} and \code{Sex}.
The Sample IDs must match the Sample IDs from the \code{gtseq} CSV file. Other metadata
columns may be present.}

\item{allele_info}{Comma separated file with information about the loci of interest.
The \code{site_id} must match the Allele IDs from the GT-seq output. the \code{advantage} column
is required if the \code{type} \code{all_alleles} is specified.}

\item{verbose}{show messages, logical - default = TRUE}
}
\value{
An RSQLite database saved to memory
}
\description{
This function takes three inputs of CSV files and transforms them into
a RSQLite database. The three inputs are: \cr
\enumerate{
\item gtseq: Output of the GTseq pipeline \cr
\item metadata: Metadata about the samples \cr
\item allele_info: Information of the loci being used \cr
}
}
\examples{
\dontrun{
DBs <- make_database(gtseq = geno, metadata = meta_data,
                     allele_info = allele_info)
}
}
