% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_all_mates.R
\name{rank_all_mates}
\alias{rank_all_mates}
\title{Rank all mates}
\usage{
rank_all_mates(females, males, ranked_list, output_head = 5)
}
\arguments{
\item{females}{Vector of sample IDs of the females to compare to the males. The Sample IDs
must match the ones found in the \code{DB}.}

\item{males}{Vector of sample IDs of the males to compare to the females. The Sample IDs
must match the ones found in the \code{DB}.}

\item{ranked_list}{List output from \code{get_all_rankings}}

\item{Output_head}{Integer value for the top number of mating pairings. Default:
top 5 mating pairings}
}
\value{
Dataframe
}
\description{
This function takes the output of \code{get_all_rankings} and ranks all possible matings so
that the sum of the score of all mate pairings is optimized. \cr
This function uses a permutation without replacement and makes a matrix of the size
of all possibilities. It is not recommended to use more than 10x10 samples.
}
\examples{
\dontrun{
all_mates <- rank_all_mates(DB = DBs, female = females,
                              males = males, type = "all_alleles",
                              bonus=NULL, weighted_alleles=NULL)
}
}
