% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_mates.R
\name{rank_mates}
\alias{rank_mates}
\title{Rank Mates}
\usage{
rank_mates(
  DB,
  female,
  males,
  type = "all_alleles",
  bonus = NULL,
  weighted_alleles = NULL
)
}
\arguments{
\item{DB}{RSQLite database generated by \code{make_database}}

\item{female}{Sample ID of the female to compare to the males. The Sample ID
must match the ones found in the \code{DB}.}

\item{males}{Sample IDs of the males to compare to a female. The Sample IDs
must match the ones found in the \code{DB}.}

\item{type}{The type of mating that is advantageous.
\enumerate{
\item \code{all_alleles}: the user specifies for each locus what is advantageous.
\item \code{assortive}: Assortive (same alleles) is advantageous for all loci
\item \code{disassortive}: Disassortive (different alleles) is advantageous for all loci
}}

\item{bonus}{Numeric value to increase mating score if the state of alleles
specified in \code{weighted_alleles} are met. If you specify .15 as your bonus,
and 2/3 of your loci match the advantageous, then .10 will be used as the bonus.
The final score will be increased by 10\%. Must specify \code{weighted_alleles}.}

\item{weighted_alleles}{vector listing allele IDs to calculate the bonus. Must
specify \code{bonus}.}
}
\value{
dataframe
}
\description{
Wrapper function to score the mating of a single female to all males. \cr
Apply this function to multiple females using the \code{get_all_rankings} function. \cr
This functions sorts the output dataframe by highest scoring male.
}
\examples{
\dontrun{
female_mates <- rank_mate(DB = DB, female = females[1],
                          males = males, type = "all_alleles",
                          bonus=NULL, weighted_alleles=NULL)
}
}
