% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_database.R
\name{save_database}
\alias{save_database}
\title{Save Database}
\usage{
save_database(DB, path)
}
\arguments{
\item{DB}{Object your database is saved in memory to.}

\item{path}{path to where you want the database saved}
}
\value{
An RSQLite database saved to disk
}
\description{
This function saves the database you made with make_database to disk
}
\examples{
\dontrun{
save_database(DB = DBs, path = "database.sqlite")
}
}
