% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mate.R
\name{score_mate}
\alias{score_mate}
\title{Score Mate}
\usage{
score_mate(DB, female, male, type, bonus = NULL, weighted_alleles = NULL)
}
\arguments{
\item{DB}{RSQLite database generated by \code{make_database}}

\item{female}{Sample ID of the female to compare to a male. The Sample ID
must match the ones found in the \code{DB}.}

\item{male}{Sample ID of the male to compare to a female. The Sample ID
must match the ones found in the \code{DB}.}

\item{type}{The type of mating that is advantageous.
\enumerate{
\item \code{all_alleles}: the user specifies for each locus what is advantageous.
\item \code{assortive}: Assortive (same alleles) is advantageous for all loci
\item \code{disassortive}: Disassortive (different alleles) is advantageous for all loci
}}

\item{bonus}{Numeric value to increase mating score if the state of alleles
specified in \code{weighted_alleles} are met. If you specify .15 as your bonus,
and 2/3 of your loci match the advantageous, then .10 will be used as the bonus.
The final score will be increased by 10\%. Must specify \code{weighted_alleles}.}

\item{weighted_alleles}{vector listing allele IDs to calculate the bonus. Must
specify \code{bonus}.}
}
\value{
dataframe
}
\description{
Score the mating of a single female and male. \cr
Apply this function to multiple matings using the \code{rank_mates} function.
}
\examples{
\dontrun{
single <- score_mate(DB = DB, female = female,
                     male = male, type = type,
                     bonus=bonus, weighted_alleles=weighted_alleles)
}
}
